requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
bcrypt==4.2.0
boto3==1.35.15
botocore==1.35.15
certifi==2024.8.30
click==8.1.7
ecdsa==0.19.0
fastapi==0.114.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
jmespath==1.0.1
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.0
pydantic==2.9.1
pydantic-settings==2.4.0
pydantic_core==2.23.3
PyMySQL==1.1.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
sqlalchemy-to-pydantic==0.0.8
starlette==0.38.5
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
